+++ test/c++/oracle8iclobinsert.C
@@ -91,7 +91,7 @@ int	main(int argc, char **argv) {
 	const char	*bindvals[5]={"4","testchar4","testvarchar4","01-JAN-2004","testlong4"};
 	const char	*subvars[4]={"var1","var2","var3",NULL};
 	const char	*subvalstrings[3]={"hi","hello","bye"};
-	int32_t		subvallongs[3]={1,2,3};
+	int64_t		subvallongs[3]={1,2,3};
 	double		subvaldoubles[3]={10.55,10.556,10.5556};
 	uint32_t	precs[3]={4,5,6};
 	uint32_t	scales[3]={2,3,4};
